S02E02

+++ llmService/ollamaService.py
@@ -4,7 +4,7 @@ from .completionService import CompletionService
 
 class OllamaService(CompletionService):
 
-    def get_completion(self, prompt: str, model: str = "gpt-4o", temperature: int = 1, response_format: str = "text") -> str:
+    def get_completion(self, prompt: str, model: str = "gpt-4o", temperature: float = 1, response_format: str = "text") -> str:
         response = ollama.chat(model='gemma2:2b', messages=[
             {
                 'role': 'system',
